+++ app/Http/Controllers/Api/V1/BootcampController.js
@@ -8,60 +8,7 @@ const geocoder = require('../../../../Utils/geocoder');
 //@route GET /api/v1/bootcamps
 //@access Public
 exports.getBootcamps = asyncHandler(async (req, res, next) => {
-  //get request query
-  let query;
-  //create query object to hold all query params
-  const reqQuery = { ...req.query };
-  //fields to exclude
-  const removeFields = ['select', 'sort', 'page', 'limit'];
-  //loop over removeFields and delete them from reqQuery
-  removeFields.forEach(param => delete reqQuery[param]);
-  //create query string
-  let queryStr = JSON.stringify(reqQuery);
-  //create operators for (gt,gte,lt,lte)
-  queryStr = queryStr.replace(/\b(gte|gt|lte|lt)\b/g, match => `$${match}`);
-  //finding resource
-  query = Bootcamp.find(JSON.parse(queryStr)).populate('courses');
-  //select fields
-  if (req.query.select) {
-    const fields = req.query.select.split(',').join(' ');
-    query = query.select(fields);
-  }
-  //sort fields
-  if(req.query.sort){
-    const sortBy = req.query.sort.split(',').join(' ');
-    query = query.sort(sortBy);
-  }else{
-    query = query.sort('-createdAt');
-  }
-  //pagination
-  const page = parseInt(req.query.page, 10) || 1;
-  const limit = parseInt(req.query.limit, 10) || 10;
-  const startIndex = (page - 1) * limit;
-  const endIndex = page * limit;
-  const total = await Bootcamp.countDocuments();
-  query = query.skip(startIndex).limit(limit);
-  const bootcamps = await query;
-  //pagination result
-  const pagination = {};
-  if (endIndex < total) {
-    pagination.next = {
-      page: page + 1,
-      limit
-    };
-  }
-  if (startIndex > 0) {
-    pagination.prev = {
-      page: page - 1,
-      limit
-    };
-  }
-  res.status(200).json({
-    success: true,
-    count: bootcamps.length,
-    pagination,
-    data: bootcamps
-  });  
+  res.status(200).json(res.advancedResults); 
 });
 
 //@desc Get bootcamp by id
@@ -69,7 +16,6 @@ exports.getBootcamps = asyncHandler(async (req, res, next) => {
 //@access Public
 exports.getBootcamp = asyncHandler(async(req, res, next) => {
   const bootcamp = await Bootcamp.findById(req.params.id);
-  //check if bootcamp exist
   if (!bootcamp) {
     return next(
       new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404)
@@ -133,7 +79,6 @@ exports.deleteBootcamp = asyncHandler(async(req, res, next) => {
 //@access Private
 exports.getBootcampsInRadius = asyncHandler(async(req, res, next) => {
   const { zipcode, distance } = req.params;
-  //Get lat/lng from geocoder
   const loc = await geocoder.geocode(zipcode);
   const lat = loc[0].latitude;
   const lng = loc[0].longitude;
